Finish Euler18
I had the problem understood wrong. I thought I was supposed to find
all the Sundays, not just the ones on the first of the month. Change the
list to be of lists and make the second element in each be a boolean of
whether that was the first of the month. No big deal.