Speed it up by 50%
I was testing two times as many numbers as necessary, so it was easy to
speed it up that way. It still takes 25 seconds on my laptop, as opposed
to the people on Project Euler who got it in 50 milliseconds on
computers from the 90's. I don't know what they're doing. C?