Try something new on Euler31
It didn't work, but I think I'm a little closer to understanding the
issue. I think that the lifetime is staying through the whole stack, so
it expects the same lifetime for every function call on the stack, which
doesn't work because the stuff higher on the stack won't last as long as
the stuff lower. I think its because it expects all the Links in the
HashSet to have the same lifetime, which they don't. But I thought I had
to pass ownership to that, and that would make the lifetime whatever it
needed to be, right? That seems logical. I dunno though.