Re-indent
Also, do some other stuff. I started on the webservice. It's not done,
and the implementation has a long way to go, but there are a mountain of
compile errors (exotic!) that might guide me a little. Also buttons that
don't work. I shouldn't be able to get that lost.