Try to speed up Euler36
I just stumbled on it and it was slow. But the Python version is really
slow and I don't know why. It's really weird. And whenever I try to do
fancy stuff it just gets worse. I guess it's just down to calling C as
much as possible and Python is slow. That's a little frustrating though.