Do a bunch of work
That was a blur. It's slower now, which doesn't make sense. We should
find out why. It should be fast. I can't figure out what's wrong. It
might be related to the fact that we're overflowing integers in the
uptree all the time. We should change those to longs. Other than that
I'm out of ideas.