It's really fast now but it doesn't work
It's at about 2.5 s on my laptop, which is pretty good, but I don't know
why. Without parallelism, it's at about 5.5, which is still way faster,
    so I don't understand that. Very weird.
But the output is way wrong and I'm not sure why. I should test with
non-parallel folds to see if that's it. But I need it to do the right
thing or else the whole thing is nonsense. So I have to fix that before
I can figure out how I made it faster and how I can make it even faster.