Make a lot of progress on WK-nominate
The first principal component is clearly party. All Republicans are
positive and all Democrats are negative (or maybe the other way around).
The second one seems to be something like bipartisanship or maybe
pro-/anti-establishment but I'm not sure. All the highest and lowest
scorers on the second component are Republicans. How weird.
I haven't even looked at the groups yet.